add generic validations
