--- src/main/webapp/util/ValidateEmail.js
@@ -1,8 +1,0 @@
-export const validateEmail = (email) => {
-  let valid = false;
-  const re = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-  if (re.test(email)) {
-    valid = true;
-  }
-  return valid;
-};
